requirement.txt
tensorflow-gpu==1.15
pycocotools==2.0.0
matplotlib==3.0.2
numpy==1.19.5
scikit-image==0.14.2
opencv-python-headless==3.4.2.17
progressbar2==3.37.1
spacy==2.3.2
keras==2.3.1
keras-transformer==0.38.0
h5py==2.10.0
yacs==0.1.8